add AskNotificationsConnected
