+++ packages/apolloschurchapp/src/ui/Onboarding/slides/AskNotifications/AskNotifications.js
@@ -1,0 +1,94 @@
+import React, { memo } from 'react';
+import PropTypes from 'prop-types';
+import {
+  PaddedView,
+  FlexedView,
+  ProgressiveImage,
+  styled,
+  H2,
+  H5,
+  Button,
+} from '@apollosproject/ui-kit';
+
+import Slide from '../../Slide';
+
+const ContentWrapper = styled({
+  height: '100%',
+})(PaddedView);
+
+const Content = styled({
+  justifyContent: 'flex-end',
+})(FlexedView);
+
+const Title = styled(({ theme }) => ({
+  color: theme.colors.primary,
+}))(H2);
+
+const StyledH5 = styled(({ theme }) => ({
+  color: theme.colors.text.secondary,
+  paddingBottom: theme.sizing.baseUnit * 1.5,
+}))(H5);
+
+// memo = sfc PureComponent 💥
+const AskNotifications = memo(
+  ({
+    imageSource,
+    slideTitle,
+    description,
+    buttonText,
+    buttonDisabled,
+    onPressButton,
+    ...props
+  }) => (
+    <Slide {...props}>
+      <ContentWrapper vertical={false}>
+        {imageSource ? (
+          <ProgressiveImage
+            source={imageSource}
+            // thumbnail={{
+            //   uri: 'https://picsum.photos/50/50/?random',
+            //   width: 50,
+            //   height: 50,
+            // }}
+            imageStyle={{ width: 320, height: 320 }}
+          />
+        ) : null}
+        <Content>
+          <Title>{slideTitle}</Title>
+          <StyledH5>{description}</StyledH5>
+          {buttonDisabled || onPressButton ? (
+            <Button
+              title={buttonText}
+              onPress={onPressButton}
+              disabled={buttonDisabled}
+              pill={false}
+            />
+          ) : null}
+        </Content>
+      </ContentWrapper>
+    </Slide>
+  )
+);
+
+AskNotifications.propTypes = {
+  /* The `Swiper` component used in `<Onboading>` looks for and hijacks the title prop of it's
+   * children. Thus we have to use a more unique name.
+   */
+  imageSource: ProgressiveImage.propTypes,
+  slideTitle: PropTypes.string,
+  description: PropTypes.string,
+  buttonText: PropTypes.string,
+  buttonDisabled: PropTypes.bool,
+  onPressButton: PropTypes.func,
+};
+
+AskNotifications.defaultProps = {
+  slideTitle: 'Can we keep you informed?',
+  description:
+    "We'll let you know when important things are happening and keep you in the loop",
+  buttonText: 'Yes, enable notifications',
+  buttonDisabled: false,
+  secondaryNavText: 'Later',
+};
+
+export default AskNotifications;
